+++ src/Config/Hint/Psr4WildcardHint.php
@@ -64,7 +64,11 @@ class Psr4WildcardHint extends AbstractHint implements WildcardHintInterface
 
             $targetClass = $this->targetPattern;
             foreach ($matches[1] as $match) {
-                $targetClass = preg_replace("/\*/", $match, $targetClass, 1);
+                $result = preg_replace("/\*/", $match, $targetClass, 1);
+
+                if ($result !== null) {
+                    $targetClass = $result;
+                }
             }
 
             if (class_exists($targetClass) || interface_exists($targetClass)) {
@@ -79,7 +83,7 @@ class Psr4WildcardHint extends AbstractHint implements WildcardHintInterface
     {
         $namespaceLength = strrpos($pattern, "\\");
 
-        return substr($pattern, 0, $namespaceLength);
+        return $namespaceLength === false ? "" : substr($pattern, 0, $namespaceLength);
     }
 
     private function validateNamespace(string $pattern, string $namespace): void